# FastRegexMatcher: use stack memory for lowercase copy of string